Move unit tests into their own folder
