+++ src/hoc/withButtonContainerLegacyMigration.js
@@ -1,7 +1,7 @@
 import { useEffect } from '@wordpress/element';
 import isBlockVersionLessThan from '../utils/check-block-version';
 import wasBlockJustInserted from '../utils/was-block-just-inserted';
-import MigrateDimensions from './migrations/dimensions';
+import MigrateDimensions from './migrations/migrateDimensions';
 
 export default ( WrappedComponent ) => {
 	return ( props ) => {
